Bit of cleanup. Moving classes into separate files.
