+++ samples/unit-testing/Version_6/Sample/Data/DiscountPolicyData.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using NServiceBus;
+
+public class DiscountPolicyData : IContainSagaData
+{
+    public Guid Id { get; set; }
+    public string Originator { get; set; }
+    public string OriginalMessageId { get; set; }
+
+    public Guid CustomerId { get; set; }
+    public decimal TotalAmount { get; set; }
+}
