+++ samples/unit-testing/Version_6/Sample/Data/MyRequest.cs
@@ -1,0 +1,3 @@
+﻿public class MyRequest
+{
+}
