+++ samples/unit-testing/Version_6/Sample/Data/MyResponse.cs
@@ -1,0 +1,3 @@
+﻿public class MyResponse
+{
+}
