+++ samples/unit-testing/Version_6/Sample/Data/SubmitOrder.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using NServiceBus;
+
+public class SubmitOrder : IMessage
+{
+    public Guid CustomerId { get; set; }
+    public Guid OrderId { get; set; }
+    public decimal TotalAmount { get; set; }
+}
