+++ samples/unit-testing/Version_6/Sample/packages.config
@@ -3,5 +3,4 @@
   <package id="NServiceBus" version="6.0.0-unstable1693" targetFramework="net452" />
   <package id="NServiceBus.Testing" version="6.0.0-unstable0051" targetFramework="net452" />
   <package id="NUnit" version="3.2.0" targetFramework="net452" />
-  <package id="NUnitTestAdapter" version="2.0.0" targetFramework="net452" />
 </packages>
